+++ src/whatsapp/abstract/abstract.repository.ts
@@ -7,6 +7,7 @@ export type IInsert = { insertCount: number };
 
 export interface IRepository {
   insert(data: any, instanceName: string, saveDb?: boolean): Promise<IInsert>;
+  update(data: any, instanceName: string, saveDb?: boolean): Promise<IInsert>;
   find(query: any): Promise<any>;
   delete(query: any, force?: boolean): Promise<any>;
 
@@ -48,6 +49,11 @@ export abstract class Repository implements IRepository {
   public insert(data: any, instanceName: string, saveDb = false): Promise<IInsert> {
     throw new Error('Method not implemented.');
   }
+
+  public update(data: any, instanceName: string, saveDb = false): Promise<IInsert> {
+    throw new Error('Method not implemented.');
+  }
+
   public find(query: any): Promise<any> {
     throw new Error('Method not implemented.');
   }